refactor action.rs, removing duplicate code to increase readability
